Undoing premature refactor. Moving delay back in to producer
